requirement.txt
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
idna==3.4
isort==5.12.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
PyExecJS==1.5.1
Pygments==2.16.1
requests==2.31.0
rich==13.5.2
six==1.16.0
urllib3==2.0.4
winregistry==1.1.1
